[FIX] product_sold_by_delivery_week: work with lower permissions
A user with basic stock and sale permissions couldn't validate a picking
because it meant writing a field into the product. Now that write is
issued with `sudo()`.

I refactored a bit the tests. They were testing very weird scenarios
that would never happen in the real world. And, what's more interesting,
they were not testing the real world scenarios.

I lowered the permissions in tests, to make sure that the module works
with the basic permissions.

Finally, I also changed some spelling mistakes that I simply could not
help but change.

@moduon MT-5114